Holding isAuthenticated in top App.js state, & passing props down using render props pattern
New component AppliedRoute takes a component and the childprops that we wish to apply to it. We can use this to give the Login component access to the userHasAuthenticated method which is home to the App component. After a successful Login, the Login component will call this.props.userHasAuthenticated(true), setting the isAuthenticated state to true at the top level. We will conditionally render the signup / login components or a logout button in the Navbar based on this state