Created two new types of Routes to render - AuthenticatedRoute & UnauthenticatedRoute
AuthenticatedRoute will check if the user is authenticated before rendering the Component (providing access). If the user is not authenticated, they are Redirected to '/login' and the URL is provided with a query parameter denoting where to redirect them back to after a successful login. UnauthenticatedRoute will see if a user is already logged in, in which case they will be Redirected to '/' after trying to access one of these routes (like /login or /signup)